refactor(contracts): invert webui + openai_compat onto product_contracts (WS5)
Both transports now compile against the product boundary instead of the
product crate: the wire DTOs they serialize, the request bodies they
deserialize, and the operation descriptor *types* they hold moved into
`ironclaw_product_contracts`, and every implementation stayed with its owner.

- `ironclaw_webui`: 228 -> 102 production `ironclaw_product::` symbols
- `ironclaw_reborn_openai_compat`: 23 -> 3 (7 -> 2 files)

New contracts modules: `descriptors`, `inbound_requests`, `product_wire`,
`workspace_views`; `admin_users` and `operator_llm` gained their wire DTOs.

The residue is structural, not shortfall: PROPOSAL §6.1.3 keeps product's
concrete command/view/capability constants in product as the frozen
inventory, and a route handler names the constant to call the surface. The
new `reborn_transport_product_boundary.rs` freezes the residue with per-entry
reasons, pins the moved vocabulary in contracts, and pins the inventory in
product so "shrinking" the residue by moving the inventory fails loudly.

Regression tests: `crates/ironclaw_architecture/tests/reborn_transport_product_boundary.rs`
(4 tests, incl. a scanner self-test and a non-vacuity assertion).
