+++ crates/ironclaw_product/src/reborn_services/projects.rs
@@ -14,192 +14,15 @@
 //! `ironclaw_filesystem::FileType`).
 
 use async_trait::async_trait;
-use chrono::{DateTime, Utc};
-use ironclaw_host_api::ids::{TenantId, UserId};
-use serde::{Deserialize, Serialize};
-use serde_json::Value as JsonValue;
 
-/// Trusted caller identity for project operations.
-///
-/// Built by the service from the authenticated caller. Never reconstructed from
-/// the request body.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ProjectCaller {
-    pub tenant_id: TenantId,
-    pub user_id: UserId,
-}
-
-/// Access role a user holds on a project. Privilege order, highest first:
-/// `Owner > Editor > Viewer` (matches the variant declaration order).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum RebornProjectRole {
-    Owner,
-    Editor,
-    Viewer,
-}
-
-/// Lifecycle state of a project.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum RebornProjectState {
-    Active,
-    Archived,
-}
-
-/// Membership grant status.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum RebornProjectMemberStatus {
-    Active,
-    Revoked,
-}
-
-/// Sanitized project view returned to the WebUI.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornProjectInfo {
-    pub project_id: String,
-    pub name: String,
-    pub description: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub icon: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub color: Option<String>,
-    /// Extensible bag (goals, GitHub links, …).
-    pub metadata: JsonValue,
-    pub state: RebornProjectState,
-    /// The calling user's effective role on this project.
-    pub role: RebornProjectRole,
-    /// RFC3339 on the wire (serde-serialized `DateTime<Utc>`); typed here to
-    /// match the other WebUI service DTOs rather than an ambiguous `String`.
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-}
-
-/// Sanitized membership grant view returned to the WebUI.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornProjectMemberInfo {
-    pub user_id: String,
-    pub role: RebornProjectRole,
-    pub status: RebornProjectMemberStatus,
-    pub granted_by: String,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-}
-
-/// Browser body for listing the caller's projects.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct RebornListProjectsRequest {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub limit: Option<u32>,
-}
-
-/// List response.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornListProjectsResponse {
-    pub projects: Vec<RebornProjectInfo>,
-}
-
-/// Single-project response (create / get / update).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornProjectResponse {
-    pub project: RebornProjectInfo,
-}
-
-/// Browser body for creating a project.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornCreateProjectRequest {
-    pub name: String,
-    #[serde(default)]
-    pub description: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub icon: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub color: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub metadata: Option<JsonValue>,
-}
-
-/// Path/body for fetching a single project.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornGetProjectRequest {
-    pub project_id: String,
-}
-
-/// Browser body for updating a project. Absent fields are left unchanged.
-///
-/// `project_id` is supplied by the route path; the handler overrides any body
-/// value, so it carries `#[serde(default)]`.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornUpdateProjectRequest {
-    #[serde(default)]
-    pub project_id: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub icon: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub color: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub metadata: Option<JsonValue>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub state: Option<RebornProjectState>,
-}
-
-/// Path/body for deleting a project.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornDeleteProjectRequest {
-    pub project_id: String,
-}
-
-/// Path/body for listing a project's members.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornListMembersRequest {
-    pub project_id: String,
-}
-
-/// Members list response.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornListMembersResponse {
-    pub members: Vec<RebornProjectMemberInfo>,
-}
-
-/// Browser body for granting a project member a role.
-///
-/// `project_id` comes from the route path (handler-overridden).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornAddMemberRequest {
-    #[serde(default)]
-    pub project_id: String,
-    pub user_id: String,
-    pub role: RebornProjectRole,
-}
-
-/// Browser body for changing a member's role.
-///
-/// `project_id` and `user_id` come from the route path (handler-overridden).
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornUpdateMemberRoleRequest {
-    #[serde(default)]
-    pub project_id: String,
-    #[serde(default)]
-    pub user_id: String,
-    pub role: RebornProjectRole,
-}
-
-/// Browser body for revoking a member.
-///
-/// `project_id` and `user_id` come from the route path (handler-overridden), so
-/// both carry `#[serde(default)]` like [`RebornUpdateMemberRoleRequest`].
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornRemoveMemberRequest {
-    #[serde(default)]
-    pub project_id: String,
-    #[serde(default)]
-    pub user_id: String,
-}
+pub use ironclaw_product_contracts::workspace_views::{
+    ProjectCaller, RebornAddMemberRequest, RebornCreateProjectRequest, RebornDeleteProjectRequest,
+    RebornGetProjectRequest, RebornListMembersRequest, RebornListMembersResponse,
+    RebornListProjectsRequest, RebornListProjectsResponse, RebornProjectInfo,
+    RebornProjectMemberInfo, RebornProjectMemberStatus, RebornProjectResponse, RebornProjectRole,
+    RebornProjectState, RebornRemoveMemberRequest, RebornUpdateMemberRoleRequest,
+    RebornUpdateProjectRequest,
+};
 
 /// Errors a project operation may produce.
 ///
