+++ crates/ironclaw_product/src/reborn_services/thread_artifact.rs
@@ -19,6 +19,8 @@ use super::{
 };
 
 pub const THREAD_ARTIFACT_SCHEMA: &str = "ironclaw.thread_artifact.v1";
+pub use ironclaw_product_contracts::product_wire::RebornThreadArtifactRequest;
+
 pub const THREAD_ARTIFACT_MAX_MESSAGES: usize = 1_000;
 const THREAD_ARTIFACT_MAX_STORED_BYTES: usize = 16 * 1024 * 1024;
 const THREAD_ARTIFACT_MAX_SERIALIZED_BYTES: usize = 20 * 1024 * 1024;
@@ -26,11 +28,6 @@ pub const THREAD_ARTIFACT_VIEW: RebornViewDescriptor = RebornViewDescriptor {
     id: "thread_artifact",
     paginated: false,
 };
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RebornThreadArtifactRequest {
-    pub thread_id: String,
-}
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct RebornThreadArtifact {
