+++ crates/ironclaw_product/src/reborn_services/views.rs
@@ -1,125 +1,22 @@
-//! Generic read conduit for descriptor-declared product views.
+//! Product-side helpers for descriptor-declared views.
+//!
+//! The typed [`ProductView`] wrapper itself moved to
+//! `ironclaw_product_contracts::descriptors` with the WS5 port inversion
+//! (PROPOSAL §6.1.3); what stays here is the parameter/page glue product's own
+//! view implementations use plus the fail-closed default provider.
+//!
+//! [`ProductView`]: ironclaw_product_contracts::descriptors::ProductView
 
 use async_trait::async_trait;
-use ironclaw_product_contracts::views::{
-    RebornViewDescriptor, RebornViewPage, RebornViewProvider, RebornViewQuery,
-};
+use ironclaw_product_contracts::views::{RebornViewDescriptor, RebornViewPage, RebornViewProvider};
 use serde::Deserialize;
 use serde::Serialize;
-use serde::de::DeserializeOwned;
-use std::marker::PhantomData;
 
 use super::{ProductSurfaceCaller, ProductSurfaceError};
 
 #[derive(Deserialize)]
 #[serde(deny_unknown_fields)]
 struct EmptyViewParams {}
-
-/// Typed declaration for one ProductSurface read view.
-///
-/// The wire conduit remains [`RebornViewQuery`] / [`RebornViewPage`]. This
-/// wrapper keeps declaration sites tied to the request/response DTOs and gives
-/// callers a shared way to encode query params and decode payloads without
-/// hand-written `serde_json` glue at every route.
-#[derive(Debug, PartialEq, Eq)]
-pub struct ProductView<Params, Output> {
-    pub id: &'static str,
-    pub paginated: bool,
-    _types: PhantomData<fn(Params) -> Output>,
-}
-
-impl<Params, Output> Clone for ProductView<Params, Output> {
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-
-impl<Params, Output> Copy for ProductView<Params, Output> {}
-
-impl<Params, Output> ProductView<Params, Output> {
-    pub const fn new(id: &'static str, paginated: bool) -> Self {
-        Self {
-            id,
-            paginated,
-            _types: PhantomData,
-        }
-    }
-
-    pub const fn paginated(id: &'static str) -> Self {
-        Self::new(id, true)
-    }
-
-    pub const fn unpaginated(id: &'static str) -> Self {
-        Self::new(id, false)
-    }
-
-    pub const fn descriptor(&self) -> RebornViewDescriptor {
-        RebornViewDescriptor {
-            id: self.id,
-            paginated: self.paginated,
-        }
-    }
-}
-
-impl<Params, Output> ProductView<Params, Output>
-where
-    Params: Serialize,
-{
-    pub fn query(
-        &self,
-        params: Params,
-        cursor: Option<String>,
-    ) -> Result<RebornViewQuery, ProductSurfaceError> {
-        Ok(RebornViewQuery {
-            view_id: self.id.to_string(),
-            params: serde_json::to_value(params).map_err(ProductSurfaceError::internal_from)?,
-            cursor,
-        })
-    }
-}
-
-impl<Params, Output> ProductView<Params, Output>
-where
-    Output: DeserializeOwned,
-{
-    pub fn decode_page(&self, page: RebornViewPage) -> Result<Output, ProductSurfaceError> {
-        serde_json::from_value(page.payload).map_err(ProductSurfaceError::internal_from)
-    }
-}
-
-impl<Params, Output> ProductView<Params, Output>
-where
-    Params: Serialize,
-    Output: DeserializeOwned,
-{
-    pub async fn query_on(
-        &self,
-        surface: &ironclaw_product_contracts::surface::BoundProductSurface,
-        params: Params,
-        cursor: Option<String>,
-    ) -> Result<Output, ProductSurfaceError> {
-        let query = self.query(params, cursor)?;
-        let page = surface
-            .query(
-                ironclaw_product_contracts::surface::ProductSurfaceQueryRequest {
-                    view_id: query.view_id,
-                    input: query.params,
-                    cursor: query.cursor,
-                    limit: None,
-                },
-            )
-            .await?;
-        let payload = page
-            .items
-            .into_iter()
-            .next()
-            .ok_or_else(ProductSurfaceError::internal)?;
-        self.decode_page(RebornViewPage {
-            payload,
-            next_cursor: page.next_cursor,
-        })
-    }
-}
 
 pub(super) fn parse_empty_view_params(
     params: serde_json::Value,
