+++ crates/ironclaw_product/tests/product_surface_inbound_contract.rs
@@ -3,9 +3,12 @@
 use base64::Engine;
 use ironclaw_host_api::ids::{AgentId, ProjectId, TenantId, ThreadId, UserId};
 use ironclaw_product::{
+    DecodeInboundAttachments, IntoProductInboundCommand, ProductInboundCommand,
+};
+use ironclaw_product_contracts::inbound_requests::{
     ProductCancelReason, ProductCancelRunRequest, ProductCreateThreadRequest,
-    ProductGateResolution, ProductInboundAttachment, ProductInboundCommand,
-    ProductResolveGateRequest, ProductRetryRunRequest, ProductSubmitTurnRequest,
+    ProductGateResolution, ProductInboundAttachment, ProductResolveGateRequest,
+    ProductRetryRunRequest, ProductSubmitTurnRequest,
 };
 use ironclaw_product_contracts::surface::{
     ProductSurfaceCaller, ProductSurfaceError, ProductSurfaceValidationCode,
