+++ crates/ironclaw_product_contracts/CLAUDE.md
@@ -149,15 +149,29 @@ shrink-only by
 
 ## Deferred by design (not missing)
 
-§6.1.3's Owns list also names the **operator service ports** (`LlmConfigService`,
-`ActiveModelReader`, `OperatorLogsService`, `OperatorServiceLifecycleService`,
-`OperatorStatusService` + their DTOs) and the command/view/capability
-**descriptor types** (`ProductSurfaceCommandDescriptor`,
-`ProductCapabilityDescriptor`, `ProductView`). Those are sourced from
-`ironclaw_product`, and CHECKLIST WS2's `operator`/`webui`/`openai_compat` flips
-own them by name. They land in this crate; they land with the PRs that repoint
-their implementors, because moving a port definition without its implementation
-buys no dependency-edge removal.
+✎ **2026-08-01, the WS5 transport inversion discharged most of this.** The
+command/view/capability **descriptor types** now live in `descriptors`; the
+inbound request bodies in `inbound_requests`; the `Reborn*` response bodies in
+`product_wire`; the operator LLM-admin DTOs in `operator_llm`; the admin-user
+wire DTOs in `admin_users`; the project/filesystem-browse DTOs in
+`workspace_views`. What is still sourced from `ironclaw_product` and owed to the
+WS5 **`operator`** row is the operator service *ports* themselves
+(`LlmConfigService`, `ActiveModelReader`, `OperatorLogsService`,
+`OperatorServiceLifecycleService`, `OperatorStatusService`) — they land with the
+PR that repoints their implementor, because moving a port definition without its
+implementation buys no dependency-edge removal.
+
+**The line that did *not* move, and must not be crossed casually.** §6.1.3 gives
+this crate the descriptor *types* and explicitly withholds "product's 27/33/18
+concrete constants, which stay in product as the frozen inventory". Those
+constants are what a route handler actually names to call the surface, so they
+are the whole reason `ironclaw_webui` (91 of them) and
+`ironclaw_reborn_openai_compat` (3) still depend on `ironclaw_product` after the
+inversion. `reborn_transport_product_boundary.rs` pins the split in **both**
+directions — the moved vocabulary must be here, and a sample of the inventory
+must **not** be — so "finish the row" cannot quietly mean "move the inventory
+too". That is an owner decision recorded on the CHECKLIST WS5 `webui` row, not a
+cleanup.
 
 `ProductCommandAdmissionService` is a special case worth recording rather than
 deciding: §6.1.3 names its "shape" for this crate, but `admit` takes a
