+++ crates/ironclaw_product_contracts/Cargo.toml
@@ -30,8 +30,10 @@ ironclaw_host_api = { path = "../ironclaw_host_api", version = "0.1.0" }
 # so the port cannot be implemented (or called) with a plain `String` that
 # `Debug`-prints itself. A value wrapper, not a framework or runtime client —
 # the contracts-purity denylist in `reborn_dependency_boundaries.rs` governs
-# the latter and this is deliberately not one.
-secrecy = "0.10"
+# the latter and this is deliberately not one. `serde` is on because
+# `RebornOperatorConfigSetRequest` and the LLM-provider upsert body carry
+# secret material on the wire and must not degrade to a `String` here.
+secrecy = { version = "0.10", features = ["serde"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 thiserror = "2"
