+++ crates/ironclaw_product_contracts/src/views.rs
@@ -8,9 +8,11 @@
 //! vocabulary, because providers legitimately sit outside product — the
 //! admin-configuration view is implemented by `ironclaw_extension_host`.
 //!
-//! Never here: any concrete view id, any provider implementation, or the
-//! typed `ProductView` declaration wrapper (that carries product's frozen
-//! request/response DTOs and stays with them).
+//! Never here: any concrete view id, or any provider implementation. The typed
+//! `ProductView` declaration wrapper lives beside the other two operation
+//! shapes in [`crate::descriptors`] — WS1.4 left it in product because it
+//! carried product-defined request/response DTOs, and the WS5 port inversion
+//! removed that reason (PROPOSAL §6.1.3 names it as a descriptor *type*).
 
 use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
