+++ crates/ironclaw_reborn_composition/src/lib.rs
@@ -194,8 +194,8 @@ pub use root::profile::{RebornCompositionProfile, RebornCompositionProfileParseE
 pub use runtime::RebornTurnDriveOutcome;
 pub use runtime::{
     AssistantReply, ConversationId, RebornRuntime, RebornRuntimeError, RebornSkillActivation,
-    RebornSkillActivationMode, RebornSkillAsset, RebornSkillBundle, RebornSkillExecutionPlan,
-    RebornSkillExecutionResult, RebornSkillSourceKind, blocked_auth_flow_canceller,
+    RebornSkillActivationMode, RebornSkillActivationSource, RebornSkillAsset, RebornSkillBundle,
+    RebornSkillExecutionPlan, RebornSkillExecutionResult, blocked_auth_flow_canceller,
     build_reborn_runtime, build_runtime, product_auth_challenge_provider,
 };
 pub use runtime_input::{
