+++ crates/ironclaw_reborn_openai_compat/CLAUDE.md
@@ -1,5 +1,18 @@
 # ironclaw_reborn_openai_compat
 
+> **Product boundary (WS5 transport inversion, 2026-08-01).** This adapter
+> speaks `ironclaw_product_contracts` — `surface` for the membrane,
+> `inbound_requests` for the bodies it constructs, `inbound`/`outbound`/
+> `projection`/`product_wire` for what crosses back — plus
+> `ironclaw_extension_contracts` for the one channel-facing enum it stamps
+> (`ProductTriggerReason`). Its whole remaining `ironclaw_product` surface is
+> three command descriptor constants (`SUBMIT_TURN_COMMAND`,
+> `CREATE_THREAD_COMMAND`, `CANCEL_RUN_COMMAND`) — product's frozen inventory
+> per PROPOSAL §6.1.3. That list is pinned exactly and shrink-only by
+> `ironclaw_architecture/tests/reborn_transport_product_boundary.rs`; a new
+> `ironclaw_product` import fails it.
+
+
 Reborn-native OpenAI-compatible API contract surface for #3283 / #4442 /
 #4443 / #4444 / #4445 / #4446 / #4447.
 ## Boundary
