+++ crates/ironclaw_reborn_openai_compat/src/ack_helpers.rs
@@ -2,8 +2,8 @@ use crate::{
     OpenAiCompatHttpError, OpenAiCompatInternalRefs, OpenAiCompatProductActionRef,
     OpenAiCompatTurnRunRef,
 };
-use ironclaw_product::ProductInboundAck;
-use ironclaw_product::RebornSubmitTurnResponse;
+use ironclaw_product_contracts::inbound::ProductInboundAck;
+use ironclaw_product_contracts::product_wire::RebornSubmitTurnResponse;
 
 pub(crate) fn internal_refs_from_ack(
     ack: &ProductInboundAck,
@@ -63,7 +63,7 @@ pub(crate) fn product_ack_from_reborn_submit(
 
 #[cfg(test)]
 mod tests {
-    use ironclaw_product::ProductInboundAck;
+    use ironclaw_product_contracts::inbound::ProductInboundAck;
     use ironclaw_turns::AcceptedMessageRef;
 
     use super::internal_refs_from_ack;
