+++ crates/ironclaw_reborn_openai_compat/src/error.rs
@@ -1,5 +1,5 @@
-use ironclaw_product::{ProductAdapterError, ProductSurfaceRejectionKind};
-use ironclaw_product::{ProductRejection, ProductRejectionKind};
+use ironclaw_host_api::product_adapter::{ProductAdapterError, ProductSurfaceRejectionKind};
+use ironclaw_product_contracts::inbound::{ProductRejection, ProductRejectionKind};
 use ironclaw_product_contracts::surface::{ProductSurfaceError, ProductSurfaceErrorCode};
 use serde::{Deserialize, Serialize};
 
@@ -476,7 +476,7 @@ fn contains_no_exposure_sentinel(value: &str) -> bool {
 
 #[cfg(test)]
 mod tests {
-    use ironclaw_product::{ProductRejection, ProductRejectionKind};
+    use ironclaw_product_contracts::inbound::{ProductRejection, ProductRejectionKind};
 
     use super::product_rejection_to_openai_error;
