+++ crates/ironclaw_webui/src/webui_v2/handlers/run_artifact.rs
@@ -1,8 +1,10 @@
 use axum::Json;
 use axum::extract::{Extension, Path, State};
 use ironclaw_product::{
-    RUN_ARTIFACT_VIEW, RebornRunArtifact, RebornRunArtifactRequest, RebornThreadArtifact,
-    RebornThreadArtifactRequest, THREAD_ARTIFACT_VIEW,
+    RUN_ARTIFACT_VIEW, RebornRunArtifact, RebornThreadArtifact, THREAD_ARTIFACT_VIEW,
+};
+use ironclaw_product_contracts::product_wire::{
+    RebornRunArtifactRequest, RebornThreadArtifactRequest,
 };
 use ironclaw_product_contracts::surface::{
     ProductSurfaceCaller, ProductSurfaceError, ProductSurfaceQueryRequest,
