+++ docs/reborn/target-architecture/CHECKLIST.md
@@ -117,9 +117,12 @@ Conventions: every code item lands with its tests and its guidance updates in th
 
 ## WS5 — Product family
 
-- [ ] `webui`: dep flips to `product_contracts`; the bearer-evidence mint import moves to `host_api`'s sealed home; gains pairing routes; verify its boundary rule updates.
+- [ ] `webui`: dep flips to `product_contracts`; the bearer-evidence mint import moves to `host_api`'s sealed home; gains pairing routes; verify its boundary rule updates. **Port-inversion half landed with the WS5 transport PR** (the mint moved with WS1.5); pairing routes and the boundary-rule check are still open. `ironclaw_webui`'s production `ironclaw_product::` usage went **228 → 102 symbols across 4 files**. Three findings the row did not predict:
+  1. **The dep cannot flip, and §6.9.4 undercounts why by 91.** §6.9.4 says webui's only non-DTO product import is the bearer-evidence mint. It is not: **91 of the 102 survivors are the concrete command/view/capability *constants*** (`THREADS_VIEW`, `SUBMIT_TURN_COMMAND`, `EXTENSION_INSTALL_CAPABILITY`, …), which §6.1.3 explicitly keeps in product as "the frozen inventory" while granting contracts only the descriptor *types*. A route handler holds the constant to call the surface, so **`webui → product` survives this row structurally**, exactly as `extension_host → product` survived WS2.1. Moving the inventory would hand the contracts crate product's surface — `reborn_transport_product_boundary.rs` therefore pins the inventory *in product* as well as pinning the moved vocabulary in contracts. **[decision]** — resolving §6.1.3-vs-§6.9.4 (move the inventory, or accept the edge and re-word §6.9.4) is an owner call, recorded not taken.
+  2. **Eleven survivors are contract-purity residue, not inventory** — `ProductAttachmentCapabilities`/`product_attachment_capabilities` (`ironclaw_attachments` budgets — the WS5 `attachments widened` row owns them), the three `ironclaw_threads`-typed thread/timeline responses, the four `ironclaw_auth`-typed extension/account responses, `RebornGetRunStateResponse` (`ironclaw_common::RunCost` + `ironclaw_loop_contracts::LoopModelUsage`), and `RebornExecuteProductCommandResponse` (`crate::commands::CommandResultView`). Same mechanical cause as WS2.1's six blocked ports: the contracts allowlist is `host_api` + `extension_contracts` and nothing else.
+  3. **The move exposed a live duplicate type name.** `RebornSkillSourceKind` existed twice — product's WebUI catalogue enum `{User, Installed, Workspace, System}` and a composition activation enum `{System, TenantShared, User}` mapped from `ironclaw_skills::SkillSourceKind`. Invisible while both sat outside contracts; §11.2.4's location scan caught it the moment the wire enum moved. Composition's is renamed `RebornSkillActivationSource` (same PR, snapshot updated). This is the class the WS8 type-audit row names for `ExtensionActivationMode`; treat the location scan as a duplicate-name detector, not only a re-export detector.
 - [ ] `operator`: implements `product_contracts` ports (ownership un-inverts; product dep dropped); route fragments move behind `host_ingress` carriers wired by composition; gains AGENTS/CLAUDE + a boundary rule (has neither today).
-- [ ] `openai_compat`: rename from `reborn_openai_compat` **[decision — severable]**; dep flips to contracts; stale `storage`/`libsql`/`postgres` feature guidance corrected in all five audited places; collapse the LibSql/Postgres ref-store newtype wrappers onto the generic fabric form (same for product's ledger wrappers).
+- [ ] `openai_compat`: rename from `reborn_openai_compat` **[decision — severable]**; dep flips to contracts; stale `storage`/`libsql`/`postgres` feature guidance corrected in all five audited places; collapse the LibSql/Postgres ref-store newtype wrappers onto the generic fabric form (same for product's ledger wrappers). **Dep-flip half landed with the WS5 transport PR:** production `ironclaw_product::` usage **23 → 3 symbols across 7 → 2 files**, and the three survivors are `SUBMIT_TURN_COMMAND` / `CREATE_THREAD_COMMAND` / `CANCEL_RUN_COMMAND` — the frozen inventory again, the same structural reason webui's dep survives. Every DTO it speaks now comes from `ironclaw_product_contracts`; it also took the `+extension_contracts` edge §6.9.3 grants, for the one channel-facing enum it stamps (`ProductTriggerReason`). Rename and ref-store collapse untouched. ✎ *Row correction:* the "stale feature guidance in five audited places" item was already discharged by the WS11.3 drift-hotfix PR (see the WS11 row's "stale feature-gating in `product`/`openai_compat`/`event_store`/`webui`/`llm`"); it is double-counted here and should be struck when this row is next edited.
 - [ ] `product` narrows: ports/DTOs out (WS1); `adapter_registry` manifest parsing → `extension_contracts`/`extension_registry` (resolving its guidance-vs-code contradiction); the ~120-symbol `host_api::product_adapter` re-export facade dissolved; slack/telegram token heuristics → packages; `external_tool_catalog` moves in from `turns`; `reborn_services` module-charter map committed (freeze ratchet stays).
 - [ ] `conversations`/`threads` naming trap fixed: rename conversations' `SessionThreadService` (→ `InboundConversationService`) + its same-named DTO trio; unify `ExternalActorRef`/`ExternalConversationRef` with `host_api` (delete product's field-by-field translators). ✎ **Amended 2026-08-01 (Wave 1 truth audit) — the counterpart crate in this row is stale, and the unification is a design decision, not a delete.** WS1.4 (#6980) moved the pair out of `host_api`, so the two declarations today are `crates/ironclaw_conversations/src/ids.rs:48,72` and **`crates/ironclaw_extension_contracts/src/external.rs:69,144`**; `ironclaw_product_contracts` only imports them (`inbound.rs:13-15`, `projection.rs:11-13`). Both are hand-written `pub struct`s. **They are not field-compatible**, so "one canonical definition, the other deleted" understates the work: conversations' actor ref is `{kind, id}` vs extension_contracts' `{kind, id, display_name}`; conversations' conversation ref is `{space_id, conversation_id, thread_id, message_id}` vs `{space_id, conversation_id, topic_id, reply_target_message_id}`; the error types differ (`InboundTurnError` vs `ProductAdapterError`). The translators are correspondingly lossy and **inconsistent with each other** — `product/src/conversation_binding.rs:818-821` silently drops `display_name`, `:826-835` maps `topic_id → thread_id` / `reply_target_message_id → message_id`, and `product/src/workflow.rs:595-606` is a *second* translator that hardcodes `None` for the fourth field, with five more inline constructions at `product/src/run_delivery/gate_routes.rs:40,48,59,68,77`. Picking the field set and the `None` semantics is the actual decision this row owns. The duplicate is already **pinned as a deliberate exemption** rather than missed — `reborn_extension_contract_location_scan.rs:120-136` lists both names in `COLLISION_EXEMPT` with the note "the same concept, declared twice … a real duplicate-surface finding, not a false positive" — so it cannot regress, but the scan will not close it either. Slot 4's finding on #6980; PROPOSAL §6.4.2 carries the matching amendment.
 - [ ] `attachments` widened: ports + composition impls move in; one home for size ceilings (webui/openai import them).
