+++ docs/reborn/target-architecture/families/contracts.md
@@ -149,8 +149,8 @@ Contracts holds the sealed constructors for the `Authorized` witness, the privil
 - **Purpose:** the neutral product-boundary vocabulary — the `ProductSurface` membrane, its caller and descriptor types, product wire DTOs, and the product-side ports whose real implementations sit beside or below product.
 - **Owns:**
   - `ProductSurface`, `BoundProductSurface`, `ProductSurfaceCaller`, and the invoke/query/stream DTOs that cross the membrane — the single generic entry point every transport (webui, the OpenAI-compatible adapter, a channel package) calls through.
-  - the command, view, and capability descriptor *types* — the shapes a concrete command or view instantiates, not the frozen inventory of concrete commands itself, which stays with product.
-  - product wire DTOs: lifecycle projections and operator menu vocabulary. ✎ **Corrected 2026-08-01 (was: "and the full event wire enumeration a transport streams to a client"):** there is no such enumeration to own. The 43-variant event enum this clause described was measured to have zero consumers outside its own crate — no transport ever streamed it, and the only live references were negative, chiefly an architecture test asserting the OpenAI-compatible routes must **not** stream it — so it was deleted rather than relocated (#6980 declined the move as an inventory correction; #6982 executed the deletion). The charter is unchanged: this crate is where product wire DTOs live. It simply has one fewer member than the spec assumed, and a reintroduction pin now guards the name (PROPOSAL §6.1.3).
+  - the command, view, and capability descriptor *types* (`ProductSurfaceCommandDescriptor`, `ProductCapabilityDescriptor`, `ProductView`) — the shapes a concrete command or view instantiates, not the frozen inventory of concrete commands itself, which stays with product. **That split is the reason a transport still names product**: a route handler holds the descriptor *constant*, so the `webui → product` dependency survives the inversion by design, not by shortfall (pinned both ways by `reborn_transport_product_boundary.rs`).
+  - product wire DTOs: lifecycle projections, operator menu vocabulary, the inbound request bodies a transport deserializes, and the `Reborn*` response bodies it serializes back. ✎ **Corrected 2026-08-01 (was: "and the full event wire enumeration a transport streams to a client"):** there is no such enumeration to own. The 43-variant event enum this clause described was measured to have zero consumers outside its own crate — no transport ever streamed it, and the only live references were negative, chiefly an architecture test asserting the OpenAI-compatible routes must **not** stream it — so it was deleted rather than relocated (#6980 declined the move as an inventory correction; #6982 executed the deletion). The charter is unchanged: this crate is where product wire DTOs live. It simply has one fewer member than the spec assumed, and a reintroduction pin now guards the name (PROPOSAL §6.1.3). ✎ *Widened 2026-08-01 (WS5 transport inversion): the `Reborn*` bodies were left in product by WS1.4; they are the reason `webui` and the OpenAI-compatible adapter compiled the product crate, and §6.1.3 already assigned them here.*
   - product-side ports whose implementations live beside product: channel delivery resolution and reply-context sourcing, command admission, the operator's LLM-config, active-model, logs, service-lifecycle, and status services, lifecycle product service vocabulary, account-connection status sourcing, and channel-config product service.
 - **Never contains:** the `ProductSurface` implementation itself; any handler, admission, or delivery logic; HTTP of any kind; projection reducers.
 - **Public surface:** the `ProductSurface` membrane and the ports above. Every port here is defined once and implemented by exactly the crate that owns the behavior — product, operator, the extension host, the extension manager, or composition — never by more than one, and never by this crate.
